Organize packages by features (#78)
* Add wired and selection packages to organize features

* Fix package change import errors
